misc: Add git-clang-format to pre-commit with wrapper script
This commit enables automatic formatting of C/C++ code using
`clang-format` via the `git-clang-format` tool. The `git-clang-format`
tool has the advantage that it only re-formats the changed lines of a
git commit, rather than full files. Therefore it can be used to
incrementally introduce automatic formatting without requiring
re-formatting of the whole repository.

This commit is intended as an optional feature, designed to help gem5
developers by automating part of the code-review process (adherence to
the gem5 style guide).

The `clang-format` tools are required to enable this feature, although
it will fail gracefully if the tools are not present.
They can be installed on Ubuntu using the command:
    `sudo apt install clang-format`

The `.clang-format` file introduced here is generated by ChatGPT to
closely match the gem5 Style Guide [1]. It is recommended that this
file is replaced by the `.clang-format` proposed in PR #362 [2] by
@trollet91 and other gem5 contributors.

This commit also introduces a wrapper script `run-git-clang-format.py`
which integrates `git-clang-format` into `pre-commit`. The wrapper
script can be run manually from the command line if desired, and has
features to allow individual files in the commit to be excluded from
the re-formatting, or to disable re-formatting completely using an
environment variable. This gives the developer control over
`clang-format` without having to disable all git hooks.

- Set the `GEM5_NO_GIT_CLANG_FORMAT` environment variable to any value
to disable `git-clang-format` completely in `pre-commit`.

- Set the `GEM5_GIT_CLANG_FORMAT_IGNORE` environment variable to a
colon-separated list of repository-root relative file paths to
disable automatic re-formatting for individual files in a commit.

[1] https://www.gem5.org/documentation/general_docs/development/coding_style/
[2] https://github.com/gem5/gem5/pull/362

Change-Id: I68c94a7b3a9b15915dc12a45f8e61baed8550806