stdlib: Add defaults for VIPER CPU cache
The CPU cache is completely bypassed when using KVM or AtomicCPU which
are the only CPUs supported by VIPER board. Move the values from config
to the stdlib code as defaults so users are not required to specify
them. This reduces the amount of code in the main x86-mi300-gpu.py file
and makes it easier to read.