misc: Rework Arm improvements
* RCpc memory model has been moved out from the highlights:
1) It is already mentioned in the ArmISA changes section
2) The contribution solely addresses the architectural interface
and does not really model the relaxedness of RCpc consistency

* FEAT_AFP has been moved out from the highlights
1) It is already mentioned in the ArmISA changes section

* Add more context to the FEAT_SVE2 entry
This is a major improvement/contribution and deserves more
details to be provided to the users

* Decouple overhaul of Arm PTW from the stdlib changes
The Arm PTW now supports multiple outstanding walks and
adopts a unified port interface. As such, there is a
relation between the former and the latter.
However, I am not sure the automatic handling of walker
ports in stdlib can be considered a major improvement. I also
believe it is better cathegorized as User-facing change
as it asks users to change their stdlib powered cache
hierarchy

Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Change-Id: Icd80474803682bde33ba35432718174c846e0872